moved test utilty functions
