+++ tests/models/affiliation_address/training_data_test.py
@@ -1,25 +1,28 @@
 import logging
-from typing import Sequence, Tuple
 
 from lxml import etree
 
 from sciencebeam_parser.document.layout_document import (
     LayoutBlock,
     LayoutDocument,
-    LayoutLine,
-    LayoutLineDescriptor
+    LayoutLine
 )
 from sciencebeam_parser.document.tei.common import get_tei_xpath_text_content_list, tei_xpath
 from sciencebeam_parser.models.data import (
-    DEFAULT_DOCUMENT_FEATURES_CONTEXT,
-    LabeledLayoutModelData,
-    LayoutModelData
+    DEFAULT_DOCUMENT_FEATURES_CONTEXT
 )
 from sciencebeam_parser.models.affiliation_address.data import AffiliationAddressDataGenerator
 from sciencebeam_parser.models.affiliation_address.training_data import (
     AffiliationAddressTeiTrainingDataGenerator
 )
 from sciencebeam_parser.utils.xml import get_text_content
+
+from tests.models.training_data_test_utils import (
+    get_labeled_model_data_list,
+    get_labeled_model_data_list_list,
+    get_model_data_list_for_layout_document,
+    get_next_layout_line_for_text
+)
 
 
 LOGGER = logging.getLogger(__name__)
@@ -39,69 +42,15 @@ def get_data_generator() -> AffiliationAddressDataGenerator:
     return AffiliationAddressDataGenerator(DEFAULT_DOCUMENT_FEATURES_CONTEXT)
 
 
-def get_model_data_list_for_layout_document(
-    layout_document: LayoutDocument
-) -> Sequence[LayoutModelData]:
-    data_generator = get_data_generator()
-    return list(data_generator.iter_model_data_for_layout_document(
-        layout_document
-    ))
-
-
-def get_labeled_model_data_list(
-    label_and_layout_line_list: Sequence[Tuple[str, LayoutLine]]
-) -> Sequence[LabeledLayoutModelData]:
-    data_generator = get_data_generator()
-    labeled_model_data_list = []
-    for label, layout_line in label_and_layout_line_list:
-        layout_document = LayoutDocument.for_blocks([LayoutBlock(lines=[layout_line])])
-        labeled_model_data_list.extend([
-            LabeledLayoutModelData.from_model_data(
-                model_data,
-                label=('B-' if index == 0 else 'I-') + label
-            )
-            for index, model_data in enumerate(
-                data_generator.iter_model_data_for_layout_document(
-                    layout_document
-                )
-            )
-        ])
-    return labeled_model_data_list
-
-
-def get_labeled_model_data_list_list(
-    label_and_layout_line_list_list: Sequence[Sequence[Tuple[str, LayoutLine]]]
-) -> Sequence[Sequence[LabeledLayoutModelData]]:
-    return [
-        get_labeled_model_data_list(label_and_layout_line_list)
-        for label_and_layout_line_list in label_and_layout_line_list_list
-    ]
-
-
-def get_layout_line_for_text(text: str, line_id: int) -> LayoutLine:
-    return LayoutLine.for_text(
-        text,
-        tail_whitespace='\n',
-        line_descriptor=LayoutLineDescriptor(line_id=line_id)
-    )
-
-
-class ModuleState:
-    line_id: int = 1
-
-
-def get_next_layout_line_for_text(text: str) -> LayoutLine:
-    line_id = ModuleState.line_id
-    ModuleState.line_id += 1
-    return get_layout_line_for_text(text, line_id=line_id)
-
-
 class TestAffiliationAddressTeiTrainingDataGenerator:
     def test_should_include_layout_document_text_in_tei_output(self):
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         layout_document = LayoutDocument.for_blocks([LayoutBlock.for_text(TEXT_1)])
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
-            get_model_data_list_for_layout_document(layout_document)
+            get_model_data_list_for_layout_document(
+                layout_document,
+                data_generator=get_data_generator()
+            )
         )
         LOGGER.debug('xml: %r', etree.tostring(xml_root))
         aff_nodes = tei_xpath(xml_root, AFFILIATION_XPATH)
@@ -115,7 +64,10 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             LayoutLine.for_text(text, tail_whitespace='\n')
         ])])
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
-            get_model_data_list_for_layout_document(layout_document)
+            get_model_data_list_for_layout_document(
+                layout_document,
+                data_generator=get_data_generator()
+            )
         )
         aff_nodes = tei_xpath(xml_root, AFFILIATION_XPATH)
         assert len(aff_nodes) == 1
@@ -128,7 +80,10 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             LayoutLine.for_text(TEXT_2, tail_whitespace='\n')
         ])])
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
-            get_model_data_list_for_layout_document(layout_document)
+            get_model_data_list_for_layout_document(
+                layout_document,
+                data_generator=get_data_generator()
+            )
         )
         aff_nodes = tei_xpath(xml_root, AFFILIATION_XPATH)
         assert len(aff_nodes) == 1
@@ -141,7 +96,10 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             LayoutLine.for_text(TEXT_2, tail_whitespace='\n')
         ])])
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
-            get_model_data_list_for_layout_document(layout_document)
+            get_model_data_list_for_layout_document(
+                layout_document,
+                data_generator=get_data_generator()
+            )
         )
         aff_nodes = tei_xpath(xml_root, AFFILIATION_XPATH)
         assert len(aff_nodes) == 1
@@ -177,7 +135,8 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             ('<institution>', get_next_layout_line_for_text(TEXT_2))
         ]
         labeled_model_data_list = get_labeled_model_data_list(
-            label_and_layout_line_list
+            label_and_layout_line_list,
+            data_generator=get_data_generator()
         )
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
@@ -213,7 +172,8 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             ('<country>', get_next_layout_line_for_text('Country 1'))
         ]
         labeled_model_data_list = get_labeled_model_data_list(
-            label_and_layout_line_list
+            label_and_layout_line_list,
+            data_generator=get_data_generator()
         )
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
@@ -263,11 +223,12 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             ('<unknown>', get_next_layout_line_for_text(TEXT_1))
         ]
         labeled_model_data_list = get_labeled_model_data_list(
-            label_and_layout_line_list
+            label_and_layout_line_list,
+            data_generator=get_data_generator()
         )
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
-            labeled_model_data_list
+            labeled_model_data_list,
         )
         LOGGER.debug('xml: %r', etree.tostring(xml_root))
         aff_nodes = tei_xpath(xml_root, AFFILIATION_XPATH)
@@ -283,7 +244,8 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             ('<institution>', get_next_layout_line_for_text(TEXT_2))
         ]
         labeled_model_data_list = get_labeled_model_data_list(
-            label_and_layout_line_list
+            label_and_layout_line_list,
+            data_generator=get_data_generator()
         )
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         xml_root = training_data_generator.get_training_tei_xml_for_model_data_iterable(
@@ -306,7 +268,8 @@ class TestAffiliationAddressTeiTrainingDataGenerator:
             ]
         ]
         labeled_model_data_list_list = get_labeled_model_data_list_list(
-            label_and_layout_line_list_list
+            label_and_layout_line_list_list,
+            data_generator=get_data_generator()
         )
         training_data_generator = AffiliationAddressTeiTrainingDataGenerator()
         xml_root = training_data_generator.get_training_tei_xml_for_multiple_model_data_iterables(
